Move proxy-window to window
