+++ src/api/crunchyroll-beta-http-client/crunchyroll-beta-http-client.ts
@@ -1,4 +1,4 @@
-import { WindowProxy } from '../../utils/proxy-window';
+import { WindowProxy } from '../../utils/window';
 import { BaseHttpClient, Config, Response } from '../base-http-client';
 import {
   GetResponseFromIndexV2,
