+++ src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './keybinds';
 export * from './message';
 export * from './page';
 export * from './player';
-export * from './proxy-window';
 export * from './redux';
 export * from './responsive';
 export * from './shadow-dom';
